Move to a pkg-private utility class.
