+++ lucene/core/src/test/org/apache/lucene/util/TestArrayUtil.java
@@ -481,117 +481,4 @@ public class TestArrayUtil extends LuceneTestCase {
 
     assertEquals(0, ArrayUtil.compareUnsigned8(a, aOffset, b, bOffset));
   }
-
-  public void testEquals4() {
-    int aOffset = TestUtil.nextInt(random(), 0, 3);
-    byte[] a = new byte[Integer.BYTES + aOffset];
-    int bOffset = TestUtil.nextInt(random(), 0, 3);
-    byte[] b = new byte[Integer.BYTES + bOffset];
-
-    for (int i = 0; i < Integer.BYTES; ++i) {
-      a[aOffset + i] = (byte) random().nextInt(1 << 8);
-    }
-    System.arraycopy(a, aOffset, b, bOffset, 4);
-
-    assertTrue(ArrayUtil.equals4(a, aOffset, b, bOffset));
-
-    for (int i = 0; i < Integer.BYTES; ++i) {
-      do {
-        b[bOffset + i] = (byte) random().nextInt(1 << 8);
-      } while (b[bOffset + i] == a[aOffset + i]);
-
-      assertFalse(ArrayUtil.equals4(a, aOffset, b, bOffset));
-
-      b[bOffset + i] = a[aOffset + i];
-    }
-  }
-
-  public void testEquals8() {
-    int aOffset = TestUtil.nextInt(random(), 0, 7);
-    byte[] a = new byte[Long.BYTES + aOffset];
-    int bOffset = TestUtil.nextInt(random(), 0, 7);
-    byte[] b = new byte[Long.BYTES + bOffset];
-
-    for (int i = 0; i < Long.BYTES; ++i) {
-      a[aOffset + i] = (byte) random().nextInt(1 << 8);
-    }
-    System.arraycopy(a, aOffset, b, bOffset, 8);
-
-    assertTrue(ArrayUtil.equals8(a, aOffset, b, bOffset));
-
-    for (int i = 0; i < Long.BYTES; ++i) {
-      do {
-        b[bOffset + i] = (byte) random().nextInt(1 << 8);
-      } while (b[bOffset + i] == a[aOffset + i]);
-
-      assertFalse(ArrayUtil.equals8(a, aOffset, b, bOffset));
-
-      b[bOffset + i] = a[aOffset + i];
-    }
-  }
-
-  public void testCommonPrefixLength4() {
-    int aOffset = TestUtil.nextInt(random(), 0, 3);
-    byte[] a = new byte[Integer.BYTES + aOffset];
-    int bOffset = TestUtil.nextInt(random(), 0, 3);
-    byte[] b = new byte[Integer.BYTES + bOffset];
-
-    for (int i = 0; i < Integer.BYTES; ++i) {
-      a[aOffset + i] = (byte) random().nextInt(1 << 8);
-      do {
-        b[bOffset + i] = (byte) random().nextInt(1 << 8);
-      } while (b[bOffset + i] == a[aOffset + i]);
-    }
-
-    for (int i = 0; i < Integer.BYTES; ++i) {
-      assertEquals(i, ArrayUtil.commonPrefixLength4(a, aOffset, b, bOffset));
-      b[bOffset + i] = a[aOffset + i];
-    }
-
-    assertEquals(4, ArrayUtil.commonPrefixLength4(a, aOffset, b, bOffset));
-  }
-
-  public void testCommonPrefixLength8() {
-    int aOffset = TestUtil.nextInt(random(), 0, 7);
-    byte[] a = new byte[Long.BYTES + aOffset];
-    int bOffset = TestUtil.nextInt(random(), 0, 7);
-    byte[] b = new byte[Long.BYTES + bOffset];
-
-    for (int i = 0; i < Long.BYTES; ++i) {
-      a[aOffset + i] = (byte) random().nextInt(1 << 8);
-      do {
-        b[bOffset + i] = (byte) random().nextInt(1 << 8);
-      } while (b[bOffset + i] == a[aOffset + i]);
-    }
-
-    for (int i = 0; i < Long.BYTES; ++i) {
-      assertEquals(i, ArrayUtil.commonPrefixLength8(a, aOffset, b, bOffset));
-      b[bOffset + i] = a[aOffset + i];
-    }
-
-    assertEquals(8, ArrayUtil.commonPrefixLength8(a, aOffset, b, bOffset));
-  }
-
-  public void testCommonPrefixLengthN() {
-    final int numBytes = TestUtil.nextInt(random(), 2, 16);
-
-    int aOffset = TestUtil.nextInt(random(), 0, numBytes - 1);
-    byte[] a = new byte[numBytes + aOffset];
-    int bOffset = TestUtil.nextInt(random(), 0, numBytes - 1);
-    byte[] b = new byte[numBytes + bOffset];
-
-    for (int i = 0; i < numBytes; ++i) {
-      a[aOffset + i] = (byte) random().nextInt(1 << 8);
-      do {
-        b[bOffset + i] = (byte) random().nextInt(1 << 8);
-      } while (b[bOffset + i] == a[aOffset + i]);
-    }
-
-    for (int i = 0; i < numBytes; ++i) {
-      assertEquals(i, ArrayUtil.commonPrefixLengthN(a, aOffset, b, bOffset, numBytes));
-      b[bOffset + i] = a[aOffset + i];
-    }
-
-    assertEquals(numBytes, ArrayUtil.commonPrefixLengthN(a, aOffset, b, bOffset, numBytes));
-  }
 }
